Rename diagnostic type files, include in netfx project
